refactor: :recycle: удалить property-механизм referenceDir
Новый full sync получает XML-ресурсы через topology и prepare-capabilities. Legacy writers и их прямые тесты удалены, а descriptors форм и макетов перенесены к topology-регистрациям.
