+++ packages/core/metadata/forms/clientApplicationForm/syncToXML.ts
@@ -5,10 +5,6 @@ import { convertClientApplicationFormFromYAMLToXML } from "./fromYAMLToXML"
 import type { ExternalMetadataContextItem } from "../../orchestration/externalMetadata/types"
 import type { ClientApplicationFormXML, ClientApplicationFormYAML, FormMetadataXML } from "./types"
 import { xmlExport } from "../../../xml/export/exporter"
-import { importFromYAML } from "../../../yaml/import"
-import importContentFromXML from "../../../xml/import/importer"
-import { copyFormItemExternalFilesToXML } from "./externalItemFiles"
-import { copyExistingRawFile, copyRawDirectoryFiles } from "./externalRawFiles"
 import type { XmlWriteManifest } from "../../orchestration/xmlWriteManifest"
 import type { PreparedYamlFile } from "../../project/preparedYamlProject"
 import type { ConfigurationIndexReader } from "../../configurationIndex/sharedSnapshot"
@@ -16,75 +12,6 @@ import { bindDeferredObjectValues, type DeferredObjectValue } from "../../orches
 import type { MetadataItemRule } from "../../orchestration/property/types"
 import { ClientApplicationFormRules } from "./rules"
 import { buildClientApplicationBaseForm } from "./baseForm"
-
-export const syncFormToXML = async (params: {
-  context: ConfigurationContextWithExportToXML
-  inputDir: string
-  formName: string
-  outputDir: string
-  referenceDir?: string
-  currentXMLPath?: string
-  xmlManifest?: XmlWriteManifest
-  rule?: MetadataItemRule
-}): Promise<void> => {
-  const { context, inputDir, formName, outputDir } = params
-  const referenceDir = params.referenceDir
-
-  const { yamlContent, formDir } = await readFormFiles({ inputDir, formName })
-
-  const yamlObj = importFromYAML<ClientApplicationFormYAML>(yamlContent)
-
-  const contextWithFormDir = createFormScopedContext({ context, formDir, currentXMLPath: params.currentXMLPath })
-  const contextWithFormExternalMetadata = createFormExternalMetadataContext({
-    context: contextWithFormDir,
-    formName,
-  })
-
-  const effectiveReferenceDir =
-    referenceDir && hasReferenceFormMetadata({ referenceDir, formName }) ? referenceDir : undefined
-  const reference = effectiveReferenceDir
-    ? readRawReferenceForm({ referenceDir: effectiveReferenceDir, formName })
-    : undefined
-  const converted = convertClientApplicationFormFromYAMLToXML({
-    context: contextWithFormExternalMetadata,
-    yaml: yamlObj,
-    name: formName,
-    referenceFormXML: reference?.formXML,
-    referenceMetadataXML: reference?.metadataXML,
-    rule: params.rule ?? ClientApplicationFormRules,
-  })
-  const isOrdinaryForm = converted.metadataXML.Form?.Properties?.FormType === "Ordinary"
-  const referenceHasFormXML = effectiveReferenceDir
-    ? hasReferenceFormXML({ referenceDir: effectiveReferenceDir, formName })
-    : true
-
-  const formXML = isOrdinaryForm && !referenceHasFormXML ? undefined : converted.formXML
-  const metadataXML = converted.metadataXML
-
-  await writeFormToXML({
-    context: contextWithFormExternalMetadata,
-    formXML,
-    metadataXML,
-    formName,
-    outputDir,
-    xmlManifest: params.xmlManifest,
-  })
-  if (formXML !== undefined) {
-    contextWithFormExternalMetadata.exportToXML.externalMetadataCollector?.recordDerived({
-      itemsTree: contextWithFormExternalMetadata.exportToXML.itemsTree,
-      segment: "Form",
-    })
-  }
-  if (formXML !== undefined) {
-    await copyFormItemExternalFilesToXML({
-      formNkdkDir: formDir,
-      formXmlDir: join(outputDir, "Forms", formName, "Ext"),
-      xmlManifest: params.xmlManifest,
-    })
-  }
-  await copyFormHelpFilesToXML({ formDir, formName, outputDir, xmlManifest: params.xmlManifest })
-  await copyFormBinToXML({ formDir, formName, outputDir, xmlManifest: params.xmlManifest })
-}
 
 export const prepareFormXML = (params: {
   context: ConfigurationContextWithExportToXML
@@ -236,20 +163,6 @@ function createFormExternalMetadataContext(params: {
   }
 }
 
-async function readFormFiles(params: { inputDir: string; formName: string }): Promise<{
-  yamlContent: string
-  formDir: string
-}> {
-  const { inputDir, formName } = params
-  const formsDir = join(inputDir, "Формы")
-  const formDir = join(formsDir, formName)
-  const yamlPath = join(formDir, "Форма.yaml")
-
-  const yamlContent = await fs.promises.readFile(yamlPath, "utf-8")
-
-  return { yamlContent, formDir }
-}
-
 const createFormScopedContext = (params: {
   context: ConfigurationContextWithExportToXML
   formDir: string
@@ -280,31 +193,6 @@ const createFormScopedContext = (params: {
   }
 }
 
-const hasReferenceFormXML = (params: { referenceDir: string; formName: string }): boolean =>
-  fs.existsSync(join(params.referenceDir, params.formName, "Ext", "Form.xml"))
-
-const hasReferenceFormMetadata = (params: { referenceDir: string; formName: string }): boolean =>
-  fs.existsSync(join(params.referenceDir, `${params.formName}.xml`))
-
-function readRawReferenceForm(params: { referenceDir: string; formName: string }): {
-  formXML?: ClientApplicationFormXML
-  metadataXML: FormMetadataXML
-} {
-  const metadataText = fs.readFileSync(join(params.referenceDir, `${params.formName}.xml`), "utf8")
-  const metadataXML = importContentFromXML<{ MetaDataObject: FormMetadataXML }>(metadataText, {
-    preserveXsiNil: true,
-    preserveEmptyElements: true,
-  }).MetaDataObject
-  const formPath = join(params.referenceDir, params.formName, "Ext", "Form.xml")
-  const formXML = fs.existsSync(formPath)
-    ? importContentFromXML<{ Form: ClientApplicationFormXML }>(fs.readFileSync(formPath, "utf8"), {
-        preserveXsiNil: true,
-        preserveEmptyElements: true,
-      }).Form
-    : undefined
-  return { metadataXML, formXML }
-}
-
 const writeFormToXML = async (params: {
   context: ConfigurationContextWithExportToXML
   formXML: ClientApplicationFormXML | undefined
@@ -329,28 +217,4 @@ const writeFormToXML = async (params: {
     await fs.promises.writeFile(formXmlPath, xmlExport({ Form: formXML }), "utf-8")
     params.xmlManifest?.addFile(formXmlPath)
   }
-}
-
-const copyFormBinToXML = async (params: {
-  formDir: string
-  formName: string
-  outputDir: string
-  xmlManifest?: XmlWriteManifest
-}): Promise<void> => {
-  const sourcePath = join(params.formDir, "Form.bin")
-  const targetPath = join(params.outputDir, "Forms", params.formName, "Ext", "Form.bin")
-  await copyExistingRawFile({ sourcePath, targetPath, xmlManifest: params.xmlManifest })
-}
-
-const copyFormHelpFilesToXML = async (params: {
-  formDir: string
-  formName: string
-  outputDir: string
-  xmlManifest?: XmlWriteManifest
-}): Promise<void> => {
-  await copyRawDirectoryFiles({
-    sourceDir: join(params.formDir, "Справка", "_files"),
-    targetDir: join(params.outputDir, "Forms", params.formName, "Ext", "Help", "_files"),
-    xmlManifest: params.xmlManifest,
-  })
 }
